inline some date/time function
optimize year hour function by using inline
