+++ be/src/column/datum.h
@@ -4,7 +4,7 @@
 
 #include <variant>
 
-#include "runtime/date_value.h"
+#include "runtime/date_value.hpp"
 #include "runtime/decimalv2_value.h"
 #include "runtime/timestamp_value.h"
 #include "storage/decimal12.h"
