+++ be/src/formats/csv/converter.h
@@ -12,7 +12,7 @@
 
 #include "common/status.h"
 #include "formats/csv/output_stream.h"
-#include "runtime/date_value.h"
+#include "runtime/date_value.hpp"
 #include "runtime/decimalv2_value.h"
 #include "runtime/timestamp_value.h"
 #include "runtime/vectorized/time_types.h"
