+++ be/src/formats/csv/output_stream.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "runtime/date_value.h"
+#include "runtime/date_value.hpp"
 #include "runtime/decimalv2_value.h"
 #include "runtime/timestamp_value.h"
