+++ be/src/storage/primary_key_encoder.cpp
@@ -33,7 +33,7 @@
 #include "column/fixed_length_column.h"
 #include "column/schema.h"
 #include "gutil/endian.h"
-#include "runtime/date_value.h"
+#include "runtime/date_value.hpp"
 #include "storage/tablet_schema.h"
 
 namespace starrocks {
