+++ be/src/storage/reader.cpp
@@ -25,7 +25,7 @@
 #include <sstream>
 #include <utility>
 
-#include "runtime/date_value.h"
+#include "runtime/date_value.hpp"
 #include "runtime/mem_pool.h"
 #include "runtime/mem_tracker.h"
 #include "runtime/string_value.hpp"
